another great readjustment of ArchSDN code structure
Because I was unhappy with the way how ArchSDN code was "scaling", I decided
  to readjust almost everything.
Datapath events are now seperated by modules, so that it is easier to read the code.
There's a big-evil module called globals.py, which ArchSDN uses to store global
access structure. If you are a C-coder, you understand the power & perils of having
global-access variables.
There's also a services module where ArchSDN services activation code is kept.

Requirements were updated:
 -> ryu is now at 4.24.

I hope this is the last time I had to remake the ArchSDN structure, but unfortunatly
developing code using the Ryu "Network Operative System" is not easy.

Signed-off-by: Carlos Miguel Ferreira <carlosmf.pt@gmail.com>