database reimplemented...
due to the stupidity of sqlite implementers, which decided to restrict the database access to just one Thread, it seems that the controller implementation using Ryu went down through the proverbial toilet.
The mix of sqlite + eventlet + pyzmq is fatal. If two greenthreads living in seperated threads decide to access the database, the interpreter will exit silently without any info.
Because sqlite has this weird implementation, I was forced to implement a lazy version of the database just in memory using python objects. It will work, but I feel that this project is no longer worthy of my attention....

Signed-off-by: Carlos Miguel Ferreira <carlosmf.pt@gmail.com>